refactor: move components to shared
