+++ src/shared/components/app-logo/index.tsx
