+++ src/shared/components/app-logo/styles.module.scss
