+++ src/shared/components/loader/index.tsx
