+++ src/shared/components/loader/styles.module.scss
