+++ src/shared/components/rss-logo/index.tsx
