+++ src/shared/components/rss-logo/styles.scss
