+++ src/widgets/header/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './styles.module.scss';
 import { HeaderMenu } from 'widgets/header-menu';
 import { UserToolbar } from 'widgets/user-toolbar';
-import { AppLogo } from 'widgets/app-logo';
+import { AppLogo } from 'shared/components/app-logo';
 import { Box } from '@mui/material';
 
 export const Header = () => {
